Use sqlite for storing the image information
We were earlier using KVariantStore, cause it is so incredibly simple to
use, but then we were just loading everything in memory and sorting it
each time. This was incredibly inefficient and it's better to store the
data properly, so that it can be efficiently queried.

ImageStorage has not been compleleted ported right now, but this commit
is rather large enough already and it's clear that some major APIs will
need to be changed.

The ImageLocationCategorizer (which actually categorized everything) has
been discarded.